Open Mobile API for packages/apps/SecureElement
This contribution is an implementation of the GlobalPlatform
Open Mobile API for Android with some modifications to
namespaces and packages to make it suitable as a core
Android component.

This contribution is based on:

0001-Open-Source-Contribution-of-Smartcard-Service-for-fr.patch
which can be found in
https://portland.source.codeaurora.org/patches/quic/la/PATCH_217881_OpenMobileAPI_20171208.tar.gz

and

0001-Code-reorganization-of-Open-Mobile-API-packages-apps.patch
which can be found in
https://portland.source.codeaurora.org/patches/quic/la/PATCH_217881_OpenMobileAPI_20171206.tar.gz

The first patch should be applied, followed by the second.

The submitted patchs were originally derived from
https://source.codeaurora.org/quic/la/platform/packages/apps/SmartCardService/commit/?h=LA.BF64.1.2.1&id=06ecea9abb8264049f52c7e31c0bc13330a425d5.

Test: Basic functionality test on target device
- Verify access rules are read from ARF on start-up
- Verify that access rules are updated on refresh tag
- Verify basic APDU transfer

Test performed in conjunction with commit
0269d238c1fc1212c8e7eb1717b47977f8f3b1fa on frameworks/base

Change-Id: Ia6c5b3657f83efa1551a067bd6aee03c68876c09
Signed-off-by: Jeremy O'Donoghue <jodonogh@codeaurora.org>