Search for rule with specific AID but for other hash
It must be possible to search the access rule cache for a rule that
contains the requested AID but with another specific hash value. Global
Platform SEAC Device Side Test Plan TC 4.1 (c0-ee-09) fails as no REF_DO
is found in the access rule cache if this fix is not applied.

Bug: 73786675
Bug: 73133863
Test: Confirmed that the TC 4.1 above can pass with this change.
Merged-In: I3965d481d07272c16c86240b901045316847719c
Change-Id: I3965d481d07272c16c86240b901045316847719c