Clear the access rule cache and the refresh tag when resetting ACE
Access Control Enforcer releases the instances of ARA controller and ARF
controller when it is reset, but the cache data of the access rules and
the refresh tag remain in the current implementation. The cache should
also be reset to avoid unexpected reuse of the cache.

Bug: 77759129
Test: Manual test to confirm that the cached is discarded as expected.

Change-Id: Id29d3c8e8118178db8c4ede915799a8129e34cb9
(cherry picked from commit cdf3a7c15cdb31c5e534df9021b720672cacf571)