Initialize Access Control Enforcer if access rules are newly installed
If no access rule is accessible, access to secure element is allowed in
eSE case and denied in UICC SE case in accordance with the chapter 4 of
GP SEAC specification. The problem with the current implementation is
that there is no mechanism to re-initialize Access Control Enforcer when
access rule becomes accessible. If access rule is installed to the
secure element, Access Controler Enforcer shall work to cache the rules.

Bug: 111195321
Test: Confirmed that new access rules are cached in the next attempt.

Change-Id: Ib56301e1150256028414d64e0fc2396e35791390